requirements.txt
Django==1.11.10
djangorestframework==3.7
psycopg2-binary==2.7.4
coreapi==2.3.3
PyYAML==3.12
openapi-codec==1.3.2
unittest-xml-reporting==1.11.0
requests==2.18.4
django-debug-toolbar==1.9.1
whitenoise==3.3.1
lxml==4.2.1
xmltodict==0.11.0
coreapi==2.3.3
django-rest-swagger==2.2.0
django-filter==1.1.0
inflect==0.2.5
